pyproject.toml
[tool.poetry]
name = "hardqode-task"
version = "0.1.0"
description = "Исполнение технического задания для компании HardQode"
authors = ["xander <moroz070688@gmail.com>"]
readme = "README.md"
packages = [{include = "hardqode_task"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.7"
django-filter = "^23.3"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
